# perf: Optimize array uniqueness handling and folder tree generation