显著优化`geneFolderTree`和`loadCollectionList`方法的性能，本地测试性能对比如下

| 方法名 | 优化前 | 优化后 |
|--------|--------|--------|
| geneFolderTree | 524.2ms | 10.91ms |
| loadCollectionList | 8523ms | 152.26ms |