@SchneeHertz 2025-09-16
这两个方法的性能开销应该集中在_.cloneDeep和_.compact(_.filter)这两部分中，我推测和_.uniq与[...new Set]关系不大，如果只更改前面的部分，优化的效果如何？