@kuviki 2025-09-16
之前的`geneFolderTree`优化有问题，我调整了，`geneFolderTree`正确的耗时应该是 10.91ms